Bugfix: raise the maximum number of steps
If we have a host with a matched unlimited virtualization subscription and N VMs
on top using it, we need at least N+1 steps to unmatch them all before attempting
to match anything different.

With a limit of 20 we should be OK in most cases.

Also tabu more entities, otherwise we might get stuck in removing and re-adding
subscriptions to those same VMs while never having the chance to remove the
subscription from the host.