Limit the amount of unaccepted moves that are generated per step
This limits the amount of time for a single step. In pathological cases,
especially small data sets, most or all of the moves might at some point
be taboo or anyway unacceptable - when this happens Optaplanner is
caught into an infinite loop. We already have a limit on the accepted
move count (acceptedCountLimit), but we don't have one for overall
moves, this patch adds it (selectedCountLimit).

This also has a beneficial effect on step time in larger problems,
because we limit the number of moves to be evaluated per step.

This of course has also a negative effect because limiting the moves
means that woese solutions will be found in large problems. To compensate
for that we allow for more steps before termination.

In a nutshell step faster, step more, avoid endless loops in corner
cases.