Optaplanner: use move tabu instead of entity tabu
In our context an entity tabu is equivalent to two moves being taboo-ed
(true -> false and false -> true).

That does not make a lot of sense, so we are transitioning to move tabu.

We size the number of tabooed moves roughly on the number the moves
we accept before termination.